+++ cli/cli.go
@@ -53,60 +53,3 @@ func CobraBindEnvironmentVariables(prefix string) func(cmd *cobra.Command, _ []s
 		})
 	}
 }
-
-// RegisterFlags registers HTTP flags with pflags
-func (c *HTTPServerConfig) RegisterFlags(flags *pflag.FlagSet, defaultPort int, defaultName string) {
-	flags.StringVarP(&c.Address, "address", "a", "localhost", "Address for server")
-	flags.IntVarP(&c.Port, "port", "p", defaultPort, "Port for server")
-	flags.StringVar(&c.Name, "server-name", defaultName, "Server Name")
-}
-
-// RegisterFlags registers Kafka flags with pflags
-func (kc *KafkaConfig) RegisterFlags(flags *pflag.FlagSet) {
-	flags.StringVarP(&kc.Broker, "kafka-broker", "b", "kafka:29092", "Kafka broker Address")
-	flags.StringVar(&kc.ClientID, "kafka-client-id", "client", "Kafka consumer Client ID")
-	flags.StringVar(&kc.TLSCaCrtPath, "kafka-server-ca-crt-path", "", "Kafka Server TLS CA Certificate Path")
-	flags.StringVar(&kc.TLSCrtPath, "kafka-client-crt-path", "", "Kafka Client TLS Certificate Path")
-	flags.StringVar(&kc.TLSKeyPath, "kafka-client-key-path", "", "Kafka Client TLS Key Path")
-	flags.BoolVar(&kc.Verbose, "kafka-verbose", false, "When this flag is set Kafka will log verbosely")
-	flags.BoolVar(&kc.JSONEnabled, "enable-json", true, "When this flag is set, messages from Kafka will be consumed as JSON instead of Avro")
-	flags.StringVar(&kc.KafkaVersion, "kafka-version", "2.1.0", "Kafka broker version")
-	flags.StringVar(&kc.ProducerCompressionCodec, "kafka-producer-compression-codec", "none", "Compression codec to use when producing messages, one of: \"none\", \"zstd\", \"snappy\", \"lz4\", \"zstd\", \"gzip\"")
-	flags.IntVar(&kc.ProducerCompressionLevel, "kafka-producer-compression-level", -1000, "Compression level to use on produced messages, -1000 signifies to use the default level.")
-	kc.SchemaRegistry = &SchemaRegistryConfig{}
-	kc.SchemaRegistry.RegisterFlags(flags)
-}
-
-// RegisterFlags register Logging flags with pflags
-func (lc *LoggingConfig) RegisterFlags(flags *pflag.FlagSet) {
-	flags.BoolVar(&lc.SentryLoggingEnabled, "sentry-logger-enabled", false, "Send error logs to Sentry")
-	flags.BoolVar(&lc.UseDevelopmentLogger, "use-development-logger", true, "Whether to use the development logger")
-	flags.StringArrayVar(&lc.OutputPaths, "log-output-paths", []string{}, "Log file path for standard logging. Logs always output to stdout.")
-	flags.StringArrayVar(&lc.ErrorOutputPaths, "log-error-output-paths", []string{}, "Log file path for error logging. Error logs always output to stderr.")
-	flags.StringVar(&lc.Level, "log-level", "info", "Application log level")
-	flags.IntVar(&lc.SamplingInitial, "log-sampling-initial", 100, "Number of log messages at given level and message to keep each second. Only valid when not using the development logger.")
-	flags.IntVar(&lc.SamplingThereafter, "log-sampling-thereafter", 100, "Keep every Nth log with a given message and threshold after log-sampling-initial is exceeded. Only valid when not using the development logger.")
-}
-
-// RegisterFlags registers Kafka flags with pflags
-func (src *SchemaRegistryConfig) RegisterFlags(flags *pflag.FlagSet) {
-	flags.StringVarP(&src.SchemaRegistryURL, "kafka-schema-registry", "r", "http://localhost:8081", "Kafka Schema Registry Address")
-}
-
-// RegisterFlags registers Sentry flags with pflags
-func (sc *SentryConfig) RegisterFlags(flags *pflag.FlagSet) {
-	flags.StringVar(&sc.DSN, "sentry-dsn", "", "Sentry DSN")
-}
-
-// RegisterFlags registers Tracer flags with pflags
-func (tc *TracingConfig) RegisterFlags(flags *pflag.FlagSet, defaultTracerName string) {
-	flags.BoolVarP(&tc.Enabled, "tracer-enabled", "t", true, "Enable tracing")
-	flags.StringVar(&tc.SamplerType, "tracer-sampler-type", "", "Tracer sampler type")
-	flags.Float64Var(&tc.SamplerParam, "tracer-sampler-param", 1.0, "Tracer sampler param")
-	flags.BoolVar(&tc.ReporterLogSpans, "tracer-reporter-log-spans", false, "Tracer Reporter Logs Spans")
-	flags.IntVar(&tc.ReporterMaxQueueSize, "tracer-reporter-max-queue-size", 100, "Tracer Reporter Max Queue Size")
-	flags.DurationVar(&tc.ReporterFlushInterval, "tracer-reporter-flush-interval", 1000000000, "Tracer Reporter Flush Interval in nanoseconds")
-	flags.StringVar(&tc.AgentHost, "tracer-agent-host", "localhost", "Tracer Agent Host")
-	flags.IntVar(&tc.AgentPort, "tracer-agent-port", 5775, "Tracer Agent Port")
-	flags.StringVar(&tc.ServiceName, "tracer-service-name", defaultTracerName, "Determines the service name for the Tracer UI")
-}
