+++ go.mod
@@ -3,12 +3,21 @@ module github.com/spothero/tools
 go 1.12
 
 require (
+	github.com/Shopify/sarama v1.22.0
+	github.com/certifi/gocertifi v0.0.0-20190415143156-92f724a62f3e // indirect
+	github.com/getsentry/raven-go v0.2.0
 	github.com/gorilla/mux v1.7.1
+	github.com/linkedin/goavro v2.1.0+incompatible
 	github.com/opentracing/opentracing-go v1.1.0
 	github.com/pkg/errors v0.8.1 // indirect
 	github.com/prometheus/client_golang v0.9.2
-	github.com/spf13/cobra v0.0.3 // indirect
+	github.com/prometheus/client_model v0.0.0-20180712105110-5c3871d89910
+	github.com/rcrowley/go-metrics v0.0.0-20181016184325-3113b8401b8a
+	github.com/spf13/cobra v0.0.3
 	github.com/spf13/pflag v1.0.3
+	github.com/spf13/viper v1.3.2
+	github.com/stretchr/objx v0.2.0 // indirect
+	github.com/stretchr/testify v1.3.0
 	github.com/uber/jaeger-client-go v2.16.0+incompatible
 	github.com/uber/jaeger-lib v2.0.0+incompatible // indirect
 	go.uber.org/atomic v1.3.2 // indirect
